+++ predictable-dots-game/src/classes/dots-game/index.ts
@@ -1,0 +1,3 @@
+export * from "./dots-game"
+export * from "./sequence-game-object"
+export * from "./dot"
