+++ predictable-dots-game/src/classes/engine/index.ts
@@ -1,4 +1,4 @@
-export * from "./GameEngine"
-export * from "./GameObject"
-export * from "./CircleGameObject"
-export * from "./AbstractGameObject"
+export * from "./game-engine"
+export * from "./game-object"
+export * from "./circle-game-object"
+export * from "./abstract-game-object"
