+++ predictable-dots-game/src/classes/index.ts
@@ -1,4 +1,2 @@
 export * from "./engine/index"
-export * from "./dot"
-export * from "./dots-game"
-export * from "./sequence-game-object"
+export * from "./dots-game/index"
